third_party/blink/web_tests/fast/js/string-localeCompare-expected.txt

This test checks String.localeCompare().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "a".localeCompare("aa") < 0 is true
PASS "a".localeCompare("b") < 0 is true
PASS "a".localeCompare("a") === 0 is true
PASS "ạ̈".localeCompare("ạ̈") === 0 is true
PASS "aa".localeCompare("a") > 0 is true
PASS "b".localeCompare("a") > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
